Multi-Lang Support
Someone once said to me I'm going to be rolled by the rest of the globe. I'm hardly the sharpest knife in the drawer. She appeared to be a little befuddled. On her forehead, she made a "L" with her finger and thumb. On the other side, the years are beginning to arrive. They didn't stop coming because they were fed up with the regulations, so I bolted. It doesn't make sense not to live for the sake of having fun. Your brain improves, while your skull deteriorates. There's so much to see and do. So, what's the harm in travelling the back roads? If you don't go, you'll never know. If you don't glow, you'll never shine.