PYTHON DOUBLE-O METHOD, HECK YEAHH
Do you ever have the impression that you're the last person on the planet? Nobody would hear a word if I screamed at the top of my lungs. I'm still on my own despite being one in a million. I'm sick of calling and waiting for someone to show up. I'm crumbling and the gaps are revealing themselves. I believe my borrowed time is running out. You might believe that you are alone. But take comfort in the fact that you aren't alone. We'll unite and refuse to let you go. You're not alone, you know.