Brainf**k Update XD
If you are unable to feed your child. Then don't have a child. And don't even think about maybe. If you are unable to feed your child. You'll be constantly trying to keep the child from crying, hustling, stealing, and lying. Now, baby is gradually dying. I said if you want to start something, you have to start something. I stated you wanted to start something. You've got to start something. It's too high to cross. It's too low to go beneath. You're trapped in the centre. And the agony is as loud as thunder. It's too high to cross. It's too low to go beneath. You're trapped in the centre. And the agony is as loud as thunder. Raise your head high. And shout to the world, "I know I'm someone!" And allow the truth to unfold. Nobody can harm you now. Because you know what's real Yes, I believe in myself, and you believe in yourself.