Security Update
I'd like to have the light without the gloom. I'd like to see the sun without the rain. I'm starting to think it's possible. To have someone that is similar to you but without the agony Will you still be here when I wake up tomorrow? I vow I'll never ask for more, but when it comes to sweetness, darling, I'm insatiable, like a kid in a candy store It's usually too wonderful, too good, too good to be true, but not tonight. I'd like to have the light without the gloom. Not tonight; I'd rather have the sun without the rain. Not tonight. I'm beginning to feel that having someone like you without the agony is conceivable. Baby.