Python3 Fix & Anti-Crash Update
There are no clouds in my stones. Let it rain, I'm going to hydroplane in the bank. The Dow Jones is falling like a stone. We'll be gone when the clouds appear. We're the Roc-A-Fellas. We fly higher than the clouds. In G5s or above You're familiar with my name (You know me) In preparation for precipitation Save your chips for a rainy day. Jay, Rain Man, has returned with Little Miss Sunshine. Rihanna, where are you? You hold my heart, and we will never be separated. Perhaps in magazines. But you'll always be my star, Baby, since you can't see sparkling automobiles in the dark. And that's when you'll need me. I'll always share with you. Because when the sun shines, we all glow. I told you I'd be here indefinitely.