Added Protection
There's a photo of our wedding day on your side of the bed. A broken clock and a bible given to you by your father It's on the window side, where the moon shines through at night. Staring at the ceiling, lying here alone, I murmured a prayer for you, then another, but you don't reach for me when you lie down softly. Tell me how you got so far away. All we have are the memories of the love we created. Are you snoozing with your own remorse? You're on your side of the bed.